chore: proper storybook implementation
