+++ src/components/reusable/codeView/blockCodeView/shidokaDarkSyntaxStyles.scss
@@ -90,6 +90,12 @@ $default-dark-font-color: #FFFFFF;
         color: #ACE1F8;
     }
 
+    .token.function_,
+    .token.class-name,
+    .hljs-title.function_ {
+        color: #F3ACD6;
+    }
+
     .hljs-section,
     .hljs-name,
     .hljs-selector-tag,
