+++ src/components/reusable/codeView/blockCodeView/shidokaLightSyntaxStyles.scss
@@ -105,8 +105,9 @@ $default-light-font-color: #3D3C3C;
         color: #236F96;
     }
 
-    .token.function,
-    .token.class-name {
+    .token.function_,
+    .token.class-name,
+    .hljs-title.function_ {
         color: #DC4B89;
     }
